requirements.txt
asgiref==3.4.1
Django==3.2.5
django-cors-headers==3.7.0
djangorestframework==3.12.4
jedi==0.18.0
parso==0.8.2
psycopg2==2.9.1
psycopg2-binary==2.9.1
Pygments==2.9.0
pytz==2021.1
sqlparse==0.4.1
urwid==2.1.2
